Forward compat of implicit scope
